# Add optimized pattern finder data structure