--- a/pkg/common/patternfinder/finder_test.go
+++ b/pkg/common/patternfinder/finder_test.go
@@ -22,6 +22,8 @@ import (
 	"strconv"
 	"strings"
 	"testing"
+
+	_ "github.com/GoogleCloudPlatform/khi/internal/testflags"
 )
 
 func TestFindAllWithStarterRunes(t *testing.T) {