… pkg/common/patternfinder/interface_test.go