Related to #255

Current KHI node parser matches its pod ID and container ID just as naive string match algorithm. It could be very heavy when a node has many containers or Pods being created in a node.

This change proposes adding `PatternFinder` interface and an implementation using Trie tree.

## Expected usage

```go
finder := patternfinder.NewTriePatternFinder[*Container]()

// step-1: add container ID patterns from logs
// example log: time="2024-01-01T01:00:00Z" level=info msg="CreateContainer within sandbox \"e4b03e280958b847e92e22b7a1570bdf63cb35432514b9a8f12f4b9adfe49714\" for &ContainerMetadata{Name:kubedns,Attempt:0,} returns container id \"eea48bce362bdf290ff0d41655c9e580a41acd354cc845c7b7163d9dd9980bd9\""

finder.AddPattern("eea48bce362bdf290ff0d41655c9e580a41acd354cc845c7b7163d9dd9980bd9", &Container{containerName:"kubedns", /* other fields...*/})
// ... Process all container creation logs and AddPatterns at first pass of processing node logs

// step-2: getting included container interfaces from log string

exampleLogBody := `time="2024-01-01T01:00:00Z" level=info msg="Stop container \"eea48bce362bdf290ff0d41655c9e580a41acd354cc845c7b7163d9dd9980bd9\" with signal terminated"`

matches := patternfinder.FindAllWithStarterRunes(finder, exampleLogBody, false, '"')
// matches[0].Value == &Container{containerName:"kubedns", /* other fields...*/}

```

## Benchmark

Increased count of pattern is negligible in Trie.
https://github.com/GoogleCloudPlatform/khi/pull/254/files#diff-78f639c007d8d1cc1acbca9dc074b2b49a95d178c5a8bbd3df6db48b3afce01fR194

```
 go test -benchmem -run=^$ -bench ^BenchmarkFindAllWithStarterRunesWithContainerIDScenario$ github.com/GoogleCloudPlatform/khi/pkg/common/patternfinder -v -count=1 -test.benchtime 10s
goos: darwin
goarch: arm64
pkg: github.com/GoogleCloudPlatform/khi/pkg/common/patternfinder
cpu: Apple M1 Pro
BenchmarkFindAllWithStarterRunesWithContainerIDScenario
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/1000container,1000query,100%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/1000container,1000query,100%hit-10                     685          17653114 ns/op         1360010 B/op       8000 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/1000container,1000query,1%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/1000container,1000query,1%hit-10                       684          17646829 ns/op         1296649 B/op       6020 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/10000container,1000query,100%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/10000container,1000query,100%hit-10                     81         148311988 ns/op         1360009 B/op       8000 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/10000container,1000query,1%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/10000container,1000query,1%hit-10                       80         147119158 ns/op         1296650 B/op       6020 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/1000container,10000query,100%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/1000container,10000query,100%hit-10                     67         177406234 ns/op        13600111 B/op      80001 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/1000container,10000query,1%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/naive/1000container,10000query,1%hit-10                       66         177117734 ns/op        12966477 B/op      60200 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/1000container,1000query,100%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/1000container,1000query,100%hit-10                     8170           1438313 ns/op          704000 B/op       3000 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/1000container,1000query,1%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/1000container,1000query,1%hit-10                      14557            807226 ns/op          640640 B/op       1020 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/10000container,1000query,100%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/10000container,1000query,100%hit-10                    6848           1683507 ns/op          704000 B/op       3000 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/10000container,1000query,1%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/10000container,1000query,1%hit-10                     13839            873640 ns/op          640640 B/op       1020 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/1000container,10000query,100%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/1000container,10000query,100%hit-10                     834          14732556 ns/op         7040002 B/op      30000 allocs/op
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/1000container,10000query,1%hit
BenchmarkFindAllWithStarterRunesWithContainerIDScenario/trie/1000container,10000query,1%hit-10                      1455           8351276 ns/op         6406401 B/op      10200 allocs/op
PASS
```