@jyane 2025-08-25
Do we know how much this can improve the situation? 

I agree with trie should be faster than the simple string comparison (as we can see the benchmark), but I wonder how much it improves in real.