@kyasbal 2025-08-25
It's hard to say without actually implementing parser using this data structure.
When the target cluster is mostly used for batch jobs, a node of the cluster may run many container and terminate them frequently compared with another cluster keeping web servers running. In those situation, I always turn off node log parser because these never ends. 

If these measurement would be necessary, I can experimentally implement the new node parser using this structure after the large refactoring ongoing now. 

However, this kind of string matcher data structure is also usable to matching resource [UID](https://kubernetes.io/docs/concepts/overview/working-with-objects/names/#uids) and IP addresses. I wonder this string match operation would be common in parsers.