@kyasbal 2025-08-26
@jyane I measured the actual time with pprof.

I tested with pprof only with 35050 count of node logs. It only consumed 5s for entire node log parser and 2.48sec was consumed for the naive log string matcher.
<img width="3360" height="3556" alt="image" src="https://github.com/user-attachments/assets/65388057-8993-44ad-9d48-2828caa6a631" />

This example log only contains 239 container on the test logs. It's quite few compared to our usual use cases on our day-to-day work. The computation complexity in current implementation is $$O(ContainerCount \times LogCount \times AvgLogStringLength)$$

$$c * 239 * 35050 = 2480ms$$ then the coefficient $$c= 0.00029605047$$ . Let's say there are 10000 containers and there are $$35050/239*10000=466527logs$$ , then the current naive algorithm would take $$0.0002960 * 10000 * 466527= 1380919.92ms = 1380s$$ but it will simply be 1.38s.


(Download the following file to see entire call tree:)
![profile001](https://github.com/user-attachments/assets/be60fd74-d06c-40ef-a1be-bfe908889a6c)